Combine code for gettling lambda functions
