+++ src/services/index.js
@@ -1,2 +1,3 @@
 export { getAWSCredentials } from "./awsCredentials";
 export { getStackResources } from "./stackResources";
+export { getLambdaFunctions } from "./awsLambda";
